LayoutTests/fast/frames/seamless/seamless-basic.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<div id="parent" style="width: 200px; height: 200px;">
<iframe id="iframe" webkitseamless src="resources/square.html"></iframe>
</div>
<script>
window.onload = function () {
    window.iframe = document.getElementById("iframe");
    shouldBeTrue("iframe.webkitseamless");

    // Seamless iframes are normal inline/replaced elements but behave
    // like blocks with (intrinsic) width filling their parent
    // and their (intrinsic) height defined by their children.

    // Spec: In visual media, in a CSS-supporting user agent: the user agent
    // should set the intrinsic width of the iframe to the width that the element
    // would have if it was a non-replaced block-level element with 'width: auto'.
    shouldBeEqualToString("window.getComputedStyle(iframe).width", "200px");

    // Spec: In visual media, in a CSS-supporting user agent: the user agent
    // should set the intrinsic height of the iframe to the height of the bounding
    // box around the content rendered in the iframe at its current width...
    shouldBeEqualToString("window.getComputedStyle(iframe).height", "100px");
}
</script>
